requirements.txt
ccxt==4.1.13
pandas==2.1.4
numpy==1.26.2
PyQt5==5.15.10
python-dotenv==1.0.0
pyinstaller==6.3.0 